# [hotfix] Encode keys as NESTED for flink keyselector