@robertwb 2019-09-03
IIRC, we were moving to simply use nested coders everywhere at the portability layer. That seems the better fix.