@sunjincheng121 2019-09-04
@robertwb  thanks a lot for the feedbacks. I agree that for making our framework as simply as possible is a good idea.

What I concerned is it will introduce some overhead for each key if we use NESTED here.
Considering the performance problem, should we make an exception here, I meant that we using OUTER for group keys case  is simple and efficient. 

What to do you think?