@sunjincheng121 2019-09-04
Hi @dmvk I have debug in my local many times, finally find this bug, and great thanks for helping double check the PreCommit test failures is unrelated  this PR.