@dmvk 2019-09-04
@kennknowles great, can you think about any `@ValidatesRunner` test case, that would ensure all runners use this correctly?