@lukecwik 2019-09-04
The proto specifically says to use NESTED:
https://github.com/apache/beam/blob/964b9647d3d77c3473539af177266506faebe1a7/model/fn-execution/src/main/proto/beam_fn_api.proto#L677

We currently use NESTED everywhere in the Beam portability APIs since NESTED vs OUTER becomes quite complicated and extremely error prone. Until there are benchmarks that show that this is a noticeable area of performance concern, we should stick with a simpler API specification.

Mental note to self is to update the proto descriptions to mention self-delimiting everywhere instead of NESTED/OUTER.