@sunjincheng121 2019-09-05
Thanks for all of your feedback! @kennknowles @robertwb @lukecwik @dmvk  

Frankly speaking, I did not do any performance testing, just an intuitive theoretical speculation. If we need to use NESTED from an architectural perspective, I agree with this decision.

I have update the PR which change the OUTER to NESTED.

Best, Jincheng