@mxm 2019-09-05
It looks to me the test issues were coming from the timer key passed from the portability layer. We could easily change the encoding context there.