@mxm 2019-09-05
Just to be clear, fixing the underlying cause is a one liner. The following line needs to be commented out: 
https://github.com/apache/beam/blob/8869fcebdb9ddff375d8e7b408f1d971e1257815/runners/flink/src/main/java/org/apache/beam/runners/flink/translation/wrappers/streaming/ExecutableStageDoFnOperator.java#L349