@tweise 2019-09-05
+1 for reverting this

I find it helpful to annotate the source, which would not only provide relevant context but also suitable author(s) to ask questions.

@sunjincheng121 please see https://beam.apache.org/contribute/
@dmvk https://beam.apache.org/contribute/committer-guide/