@sunjincheng121 2019-09-06
Thank you very much for all of your feedback! 

There are three solutions to this problem in my thoughts (we have already mentioned it above):

  **Approach 1:** Use OUTER in both keyselector and portability level.
     Benefit: no need to add length to the key, very simple solution
     Disadvantages: The architecture does not meet Luke's reference to the "use NESTED everywhere in the Beam portability" principle.

  **Approach 2:** NESTED is used in both keselector and portability level.
     Benefits: In line with Luke's reference to the "use NESTED everywhere in the Beam portability" principle.
     Disadvantages: Added length on the key(but there are currently no clear performance issues found).

  **Approach 3:** Use OUTER in the keyselector and NESTED in the portability level.
    Benefits: small code modification
    Disadvantages:
      - Decode the key every time when read/write/clean the state, so the performance is not prefect.

I have did a simple test for `prepareStateBackend`:
case1:
  **1000 characters** in length, loop 10000 times, call prepareStateBackend
  With decode: Time used: 84ms
  Without decode: Time used: 46ms

case2:
 **10000 characters** in length, loop 10000 times, call prepareStateBackend
 With decode: Time used: 227ms
 Without decode: Time used: 116ms

Hi @mxm,  regarding you said: "break backwards-compatibility.". Do you meant that the break state backwards-compatibility,  i.e.,  when change the OUTER to NESTED, the key-group-id will be changed, then user can not get the old state when upgrade the job. If so, #3 also encounter same problem. because we chanege the current key of state, user also cannot get the old state when upgrade the job. If I understand correctly, all of the approaches should have the backwards-compatibility problem.  So, please correct me If there anything I misunderstand.

So, for now, I still prefer #2.

What do you think?  @kennknowles @robertwb @dmvk @mxm @lukecwik @tweise 

Best,
Jincheng