@dmvk 2019-09-06
Personally. I lean towards 2) as well, because it's way more error prone. These things are super hard to test and it makes the runner code hard to understand (at least for me, because I'm not that familiar with the runner code-base).

Also coder contexts are marked as deprecated, is the deprecation still valid?