@tweise 2019-09-06
@sunjincheng121 thanks for the follow-up. I think this discussion would be worth having on the mailing list though :)