@sunjincheng121 2019-09-06
Thanks for the reply @dmvk @lukecwik @tweise 
I think It's good idea for bring up the discussion in mailling list @tweise 
I would like to do it and let's try to find out the final solution for this problem as soon as possible.(After all, @dmvk 's PR has been opened for a long time.）