@tweise 2019-09-06
@sunjincheng121 that sounds good.

There is another PR that will unblock the 1.9 runner: https://github.com/apache/beam/pull/9484

The 1.9 runner PR has open issues on its own also.