@mxm 2019-09-06
>(After all, @dmvk 's PR has been opened for a long time.）

It may have been open for a long time but it was only considered ready-to-review with tests passing yesterday. I too want this to be part of 1.9 but we can't force it. Things need enough time to be reviewed to ensure the quality of the next Beam release.