@mxm 2019-09-06
>I can see you already do more effort for that PR. 

I don't consider a response time of one day to be long. Please keep in mind that people may be occupied with other tasks.